test: refine permanent area vertex linking coverage
